Move FabricBlock to own class
