+++ src/main/java/paragon/minecraft/wilytextiles/blocks/AxialMultipleBlock.java
@@ -11,9 +11,6 @@ import net.minecraft.state.StateContainer;
 import net.minecraft.util.Direction;
 import net.minecraft.util.Direction.Axis;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.shapes.ISelectionContext;
-import net.minecraft.util.math.shapes.VoxelShape;
-import net.minecraft.world.IBlockReader;
 import net.minecraft.world.IWorldReader;
 
 /**
@@ -29,15 +26,6 @@ public class AxialMultipleBlock extends Block {
 	public static final IntegerProperty COUNT = IntegerProperty.create("count", MIN_COUNT, MAX_COUNT);
 	public static final EnumProperty<Direction.Axis> FACING = EnumProperty.create("facing", Direction.Axis.class, Direction.Axis.X, Direction.Axis.Z);
 
-	private static final int SHAPE_INSET = 2;
-	public static final VoxelShape SHAPE_X_1 = Block.makeCuboidShape(0 + SHAPE_INSET, 0, 0, 16 - SHAPE_INSET, 16.0 / 3.0, 16);
-	public static final VoxelShape SHAPE_X_2 = Block.makeCuboidShape(0 + SHAPE_INSET, 0, 0, 16 - SHAPE_INSET, 32.0 / 3.0, 16);
-	public static final VoxelShape SHAPE_X_3 = Block.makeCuboidShape(0 + SHAPE_INSET, 0, 0, 16 - SHAPE_INSET, 16, 16);
-	
-	public static final VoxelShape SHAPE_Z_1 = Block.makeCuboidShape(0, 0, 0 + SHAPE_INSET, 16, 16.0 / 3.0, 16 - SHAPE_INSET);
-	public static final VoxelShape SHAPE_Z_2 = Block.makeCuboidShape(0, 0, 0 + SHAPE_INSET, 16, 32.0 / 3.0, 16 - SHAPE_INSET);
-	public static final VoxelShape SHAPE_Z_3 = Block.makeCuboidShape(0, 0, 0 + SHAPE_INSET, 16, 16, 16 - SHAPE_INSET);
-
 	public AxialMultipleBlock(Properties properties) {
 		super(properties);
 		this.setDefaultState(this.stateContainer.getBaseState().with(AxialMultipleBlock.FACING, Direction.Axis.X).with(AxialMultipleBlock.COUNT, AxialMultipleBlock.MIN_COUNT));
@@ -51,23 +39,6 @@ public class AxialMultipleBlock extends Block {
 		BlockState state = context.getWorld().getBlockState(context.getPos());
 		Axis axis = context.getPlacementHorizontalFacing().getAxis();
 		return state.isIn(this) ? AxialMultipleBlock.incrementCount(state) : AxialMultipleBlock.withAxis(this.getDefaultState(), axis);
-	}
-
-	@Override
-	public VoxelShape getShape(BlockState state, IBlockReader world, BlockPos position, ISelectionContext context) {
-		int count = AxialMultipleBlock.getCountFrom(state);
-		Axis axis = AxialMultipleBlock.getAxisFrom(state);
-		switch (count) {
-			case 9:
-			case 8:
-			case 7:
-				return axis.equals(Axis.X) ? SHAPE_X_3 : SHAPE_Z_3;
-			case 6:
-			case 5:
-			case 4:
-				return axis.equals(Axis.X) ? SHAPE_X_2 : SHAPE_Z_2;
-		}
-		return axis.equals(Axis.X) ? SHAPE_X_1 : SHAPE_Z_1;
 	}
 
 	@Override
